requirements.txt
pandas
torch
transformers==4.44.2
weaviate-client
langchain
langchain-community
langchain-weaviate
sentence-transformers
transformers
datasets
accelerate
python-dotenv
tiktoken
fastapi
uvicorn[standard]
gradio
cohere
FlagEmbedding
langgraph
ollama
langchain_ollama
python-dotenv
pyevall
evaluate
nltk